+++ apps/web/src/lib/attachments/index.ts
@@ -1,0 +1,1 @@
+export { AttachmentFactory } from './AttachmentFactory'
